requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-environ
django-cors-headers
mysqlclient
celery
redis
Pillow
django-simple-captcha
pytest
pytest-django
django-cors-headers
django-simple-captcha
twilio